core/core.py
class ASIMainControlLoop:
    def __init__(self):
        # Initialize the ASI system
        self.state = {}
        self.initialize_system()

    def initialize_system(self):
        # Placeholder for system initialization logic
        pass

    def process_input(self, input_data):
        # Placeholder for input processing logic
        pass

    def make_decision(self):
        # Placeholder for decision-making logic
        pass

    def execute_action(self, action):
        # Placeholder for action execution logic
        pass

    def run(self):
        # Main control loop
        while True:
            input_data = self.get_input()
            self.process_input(input_data)
            decision = self.make_decision()
            self.execute_action(decision)

    def get_input(self):
        # Placeholder for getting input data
        return {}

# Example usage
if __name__ == "__main__":
    asi = ASIMainControlLoop()
    asi.run()
