tests/test_core.py
import unittest
from core.core import ASIMainControlLoop

class TestASIMainControlLoop(unittest.TestCase):
    def setUp(self):
        self.asi = ASIMainControlLoop()

    def test_initialization(self):
        self.assertIsInstance(self.asi, ASIMainControlLoop)
        self.assertIsInstance(self.asi.state, dict)

    def test_process_input(self):
        input_data = {"key": "value"}
        self.asi.process_input(input_data)
        # Add assertions based on the expected behavior of process_input

    def test_make_decision(self):
        decision = self.asi.make_decision()
        # Add assertions based on the expected behavior of make_decision

    def test_execute_action(self):
        action = {"action": "test"}
        self.asi.execute_action(action)
        # Add assertions based on the expected behavior of execute_action

    def test_run(self):
        # This test will need to be adapted once the run method is fully implemented
        pass

if __name__ == "__main__":
    unittest.main()
